pyproject.toml
[project]
name = "tds-project-1-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "openai>=2.4.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
